+++ packages/spor-react/src/input/Combobox.tsx
@@ -21,8 +21,8 @@ export type ComboboxProps<T> = Exclude<
     inputRef?: React.RefObject<HTMLInputElement>;
     /** If you want to allow an empty collection */
     allowsEmptyCollection?: boolean;
-    lefticon?: ReactNode;
-    righticon?: ReactNode;
+    leftIcon?: ReactNode;
+    rightIcon?: ReactNode;
     variant?: "core" | "floating";
     children?: React.ReactNode;
   };
@@ -54,8 +54,8 @@ export const Combobox = (props: ComboboxProps<object>) => {
   const {
     label,
     loading,
-    lefticon,
-    righticon,
+    leftIcon,
+    rightIcon,
     borderBottomLeftRadius = "sm",
     borderBottomRightRadius = "sm",
     borderTopLeftRadius = "sm",
@@ -109,7 +109,7 @@ export const Combobox = (props: ComboboxProps<object>) => {
     paddingLeft,
     paddingX,
     paddingY,
-    lefticon,
+    leftIcon,
   };
 
   const {
@@ -150,6 +150,7 @@ export const Combobox = (props: ComboboxProps<object>) => {
         }
         _active={{ backgroundColor: "core.surface.active" }}
         {...inputProps}
+        startElement={leftIcon}
         endElement={
           loading ? (
             <ColorSpinner
@@ -164,7 +165,7 @@ export const Combobox = (props: ComboboxProps<object>) => {
               }}
             />
           ) : (
-            righticon
+            rightIcon
           )
         }
         placeholder=""